Revise sample config: PCB line with AMR/Assembler/Dexterous Arm
- Rename robots: AMR ($40k, 1.0 m/s), Stationary Assembler ($25k),
  Dexterous Robotic Arm ($80k)
- 3 PCB stations: Component Placement, Solder Reflow, Quality Inspection
- Buffer size 10, distance 60m — single AMR creates 120-tick NVA overhead
- 50-part job demonstrates recurring fetch/deliver starvation bottleneck
- Fix csv_parser: parse robot speed as float (was int, broke 1.0 m/s)

+++ simulation/csv_parser.py
@@ -130,7 +130,7 @@ def parse_csv(text: str) -> SimConfig:
         if len(row1) < 3:
             raise ParseError(f"[ROBOT_TYPES] line {i+1}: expected type_name,speed,cost")
         type_name = row1[0]
-        speed = int(row1[1]) if row1[1] else None
+        speed = float(row1[1]) if row1[1] else None
         cost = int(row1[2])
 
         i += 1
